pyproject.toml
[project]
name = "jif-tube-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.115.12",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "rich>=14.0.0",
    "sqlmodel>=0.0.24",
]
